requirements.txt
Django==3.2.8
django-cors-headers==3.13.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1